fix: refactor display recent projects
