+++ src/renderer/components/_atoms/index.ts
@@ -2,8 +2,8 @@ export * from './buttons'
 export * from './card'
 export { File } from './file'
 export { InputWithRef } from './input'
+export * from './label'
 export * from './select'
 export { Tab } from './tab'
 export { TabList } from './tab-list'
 export * from './table'
-
