+++ src/renderer/components/_atoms/label/index.tsx
@@ -2,12 +2,12 @@ import { cn } from '@root/utils'
 import { ComponentPropsWithoutRef } from 'react'
 
 type LabelProps = ComponentPropsWithoutRef<'label'> & {
-  label: string
+  label?: string | undefined
 }
 const Label = (props: LabelProps) => {
   const { label, className, ...res } = props
   return (
-    <span className={cn('text-neutral-1000', className)} {...res}>
+    <span className={cn('text-white', className)} {...res}>
       {label}
     </span>
   )
