+++ src/renderer/components/_molecules/file/file-shape.tsx
