+++ src/renderer/components/_organisms/display-recent-projects/index.tsx
