+++ src/renderer/components/elements/index.ts
@@ -1,2 +1,2 @@
+export { default as FileElement } from '../_molecules/file'
 export { default as Dropdown } from './dropdown'
-export { default as FileElement } from './file'
