+++ src/renderer/screens/start-screen.tsx
@@ -6,9 +6,9 @@ import { FolderIcon, PlusIcon, StickArrowIcon, VideoIcon } from '../assets'
 import { useToast } from '../components/_features/[app]/toast/use-toast'
 import { MenuDivider, MenuItem, MenuRoot, MenuSection } from '../components/_features/[start]/menu'
 import { ProjectModal } from '../components/_features/[start]/new-project/project-modal'
+import DisplayRecentProjects from '../components/_organisms/display-recent-projects'
 import { ProjectFilterBar } from '../components/_organisms/project-filter-bar'
 import { StartMainContent, StartSideContent } from '../components/_templates'
-import DisplayRecentProjects from '../components/refactor/display-recent-projects'
 import { useOpenPLCStore } from '../store'
 import { FlowType } from '../store/slices/flow/types'
